文件下载，金额为0.01冗余，

+++ app/src/main/java/cn/lex_mung/client_android/mvp/model/entity/other/WebGoPayEntity.java
@@ -14,10 +14,10 @@ public class WebGoPayEntity {
     private int requireTypeId;
     private String requireTypeName;
 
-    public int getMoney() {
+    public float getMoney() {
         if(TextUtils.isEmpty(money))
             return 0;
-        return Integer.valueOf(money);
+        return Float.valueOf(money);
     }
 
     public void setMoney(String money) {
